refactor(agents): architecture pass across 6 candidates
Six surgical architecture changes shipped together:

* Add `sanitizeAndNormalizeJson` helper in `core/sanitizer.ts` and replace
  four inline `parse → normalizeForVault → redactSecretLiterals → stringify`
  sites (cursor MCP, vscode MCP, claude marketplace, claude plugin hook).
* Make `AgentDefinition.snapshot/apply` take `AgentSyncConfig` so options
  are first-class instead of duct-taped through `withClaudeOptions`
  wrappers duplicated in push and pull. Only Claude reads from the
  config today (`config.claudePlugins.syncMarketplace`); the other four
  adapters accept it via `_config?`. `ClaudeSyncOptions` deleted.
* Extract Claude plugin-apply functions and the `applyClaudePluginsDir`
  dispatcher into `agents/claude-plugin-apply.ts`. `claude.ts` shrinks
  from 774 to 549 lines.
* Move the four Claude-specific sanitizers (`sanitizeClaudeHooks/Mcp`,
  `sanitizeClaudePluginManifest/Mcp`) out of `core/sanitizer.ts` into
  `agents/claude-sanitize.ts`. `core/` now strictly holds cross-cutting
  primitives. Tests move alongside to
  `agents/__tests__/claude-sanitize.test.ts`.
* Add `defineTranslator` in `migrate/types.ts` to centralise the
  trim+empty guard. Refactor ~20 base translators across global-rules,
  commands, rules, and skills.
* Update three `_frontmatter` import paths to `frontmatter` and add
  module-level JSDoc clarifying the contract.

Also: extract `ensureCommandBackup` into `agents/_utils.ts` so
`claude-plugin-apply.ts` no longer imports back from `claude.ts`
(removes a latent ESM circular import).

Verified: `bun run check` passes — 608 tests, lint clean, typecheck
clean. Two senior-review iterations resolved before commit.
