+++ src/agents/claude.ts
@@ -1,43 +1,36 @@
-import { mkdir, readdir, readFile, stat, writeFile } from "node:fs/promises";
+import { mkdir, readdir, readFile } from "node:fs/promises";
 import { homedir } from "node:os";
 import { join } from "node:path";
 import { log } from "@clack/prompts";
 import { AgentPaths } from "../config/paths";
-import {
-  denormalizeFromVault,
-  denormalizeStringFromVault,
-  normalizeForVault,
-} from "../core/path-portability";
-import {
-  redactSecretLiterals,
-  sanitizeClaudeHooks,
-  sanitizeClaudeMcp,
-  sanitizeClaudePluginManifest,
-  sanitizeClaudePluginMcp,
-  shouldNeverSync,
-} from "../core/sanitizer";
+import type { AgentSyncConfig } from "../config/schema";
+import { denormalizeFromVault, denormalizeStringFromVault } from "../core/path-portability";
+import { sanitizeAndNormalizeJson, shouldNeverSync } from "../core/sanitizer";
 import { extractArchive } from "../core/tar";
 import {
   atomicWrite,
   collect,
+  ensureCommandBackup,
   readIfExists,
   type SnapshotArtifact,
   type SnapshotResult,
 } from "./_utils";
-import { collectClaudePlugins, InvalidPluginNameError, validatePluginName } from "./claude-plugins";
+import { applyClaudePluginsDir } from "./claude-plugin-apply";
+import { collectClaudePlugins } from "./claude-plugins";
+import {
+  sanitizeClaudeHooks,
+  sanitizeClaudeMcp,
+  sanitizeClaudePluginManifest,
+  sanitizeClaudePluginMcp,
+} from "./claude-sanitize";
 import { collectSkillArtifacts, InvalidSkillNameError, validateSkillName } from "./skills-walker";
-
-/** Options that gate optional Claude surfaces (the marketplace catalog today). */
-export interface ClaudeSyncOptions {
-  /** Sync `~/.claude/.claude-plugin/marketplace.json` when the user opts in. */
-  syncMarketplace?: boolean;
-}
 
 /** Snapshot payload for the Claude adapter. */
 export type ClaudeSnapshotResult = SnapshotResult;
 
 /** Collect Claude files that are safe to store in the encrypted vault. */
-export async function snapshotClaude(options: ClaudeSyncOptions = {}): Promise<SnapshotResult> {
+export async function snapshotClaude(config: AgentSyncConfig): Promise<SnapshotResult> {
+  const syncMarketplace = config.claudePlugins?.syncMarketplace ?? false;
   const artifacts: SnapshotArtifact[] = [];
   const warnings: string[] = [];
 
@@ -214,20 +207,18 @@ export async function snapshotClaude(options: ClaudeSyncOptions = {}): Promise<S
   // Off by default — teams have to opt in via `claudePlugins.syncMarketplace`
   // because the catalog can pin third-party sources that not every team wants
   // to standardize on through the vault.
-  if (options.syncMarketplace === true) {
+  if (syncMarketplace) {
     const marketplaceRaw = await readIfExists(AgentPaths.claude.marketplaceJson);
     if (marketplaceRaw !== null && !shouldNeverSync(AgentPaths.claude.marketplaceJson)) {
       try {
-        const parsed = JSON.parse(marketplaceRaw) as unknown;
-        const normalized = normalizeForVault(parsed, homedir());
-        const redacted = redactSecretLiterals(normalized, "marketplace");
+        const sanitized = sanitizeAndNormalizeJson(marketplaceRaw, "marketplace");
         artifacts.push({
           vaultPath: "claude/marketplace.json.age",
           sourcePath: AgentPaths.claude.marketplaceJson,
-          plaintext: `${JSON.stringify(redacted.value, null, 2)}\n`,
-          warnings: redacted.warnings,
+          plaintext: sanitized.value,
+          warnings: sanitized.warnings,
         });
-        warnings.push(...redacted.warnings);
+        warnings.push(...sanitized.warnings);
       } catch (err) {
         warnings.push(
           `[claude] Skipping marketplace.json — invalid JSON: ${
@@ -304,19 +295,14 @@ async function collectPluginHooksDir(
       continue;
     }
     try {
-      const parsed = JSON.parse(raw) as unknown;
-      // Plugin hooks routinely reference commands or working directories
-      // under HOME, so they need the same normalize-before-redact treatment
-      // as plugin.json and .mcp.json to round-trip across machines.
-      const normalized = normalizeForVault(parsed, homedir());
-      const redacted = redactSecretLiterals(normalized, "pluginHook");
+      const sanitized = sanitizeAndNormalizeJson(raw, "pluginHook");
       artifacts.push({
         vaultPath: `${vaultPrefix}/${name}.age`,
         sourcePath,
-        plaintext: `${JSON.stringify(redacted.value, null, 2)}\n`,
-        warnings: redacted.warnings,
+        plaintext: sanitized.value,
+        warnings: sanitized.warnings,
       });
-      warnings.push(...redacted.warnings);
+      warnings.push(...sanitized.warnings);
     } catch (err) {
       warnings.push(
         `[claude] Skipping plugin hook ${sourcePath} — invalid JSON: ${
@@ -374,16 +360,6 @@ export async function applyClaudeMcp(claudeJsonContent: string): Promise<void> {
   await atomicWrite(AgentPaths.claude.mcpJson, `${JSON.stringify(existing, null, 2)}\n`);
 }
 
-/** Preserve the previous command or agent file before overwrite during apply. */
-export async function ensureCommandBackup(path: string): Promise<void> {
-  try {
-    await stat(path);
-    await writeFile(`${path}.bak`, await readFile(path, "utf8"), "utf8");
-  } catch {
-    // No existing file to backup.
-  }
-}
-
 /** Restore one Claude command markdown file from the vault. */
 export async function applyClaudeCommand(commandName: string, content: string): Promise<void> {
   const target = join(AgentPaths.claude.commandsDir, commandName);
@@ -409,93 +385,6 @@ export async function applyClaudeRule(ruleName: string, content: string): Promis
 }
 
 // ─── Claude plugin apply helpers (issue #31) ─────────────────────────────────
-
-/**
- * Build the absolute plugin root path after validating the plugin name.
- * Centralised so every plugin apply helper goes through the same trust check
- * before any filesystem write.
- */
-function pluginRootFor(pluginName: string): string {
-  validatePluginName(pluginName);
-  return join(AgentPaths.claude.pluginsDir, pluginName);
-}
-
-/** Restore one plugin's `.claude-plugin/plugin.json` manifest. */
-export async function applyClaudePluginManifest(
-  pluginName: string,
-  content: string,
-): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  const target = join(root, ".claude-plugin", "plugin.json");
-  await mkdir(join(root, ".claude-plugin"), { recursive: true });
-  await ensureCommandBackup(target);
-  await atomicWrite(target, denormalizeStringFromVault(content, homedir()));
-}
-
-/** Restore one plugin-local command markdown file. */
-export async function applyClaudePluginCommand(
-  pluginName: string,
-  fileName: string,
-  content: string,
-): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  const dir = join(root, "commands");
-  const target = join(dir, fileName);
-  await mkdir(dir, { recursive: true });
-  await ensureCommandBackup(target);
-  await atomicWrite(target, content);
-}
-
-/** Restore one plugin-local agent markdown file. */
-export async function applyClaudePluginAgent(
-  pluginName: string,
-  fileName: string,
-  content: string,
-): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  const dir = join(root, "agents");
-  const target = join(dir, fileName);
-  await mkdir(dir, { recursive: true });
-  await ensureCommandBackup(target);
-  await atomicWrite(target, content);
-}
-
-/** Restore one plugin-local hook bundle JSON file. */
-export async function applyClaudePluginHook(
-  pluginName: string,
-  fileName: string,
-  content: string,
-): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  const dir = join(root, "hooks");
-  const target = join(dir, fileName);
-  await mkdir(dir, { recursive: true });
-  await ensureCommandBackup(target);
-  await atomicWrite(target, denormalizeStringFromVault(content, homedir()));
-}
-
-/** Restore one plugin's `.mcp.json`. */
-export async function applyClaudePluginMcp(pluginName: string, content: string): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  const target = join(root, ".mcp.json");
-  await mkdir(root, { recursive: true });
-  await ensureCommandBackup(target);
-  await atomicWrite(target, denormalizeStringFromVault(content, homedir()));
-}
-
-/** Restore one plugin-local skill tar archive into the plugin's skills dir. */
-export async function applyClaudePluginSkill(
-  pluginName: string,
-  skillName: string,
-  base64Tar: string,
-): Promise<void> {
-  const root = pluginRootFor(pluginName);
-  validateSkillName(skillName);
-  const targetDir = join(root, "skills", skillName);
-  await mkdir(targetDir, { recursive: true });
-  const tarBuffer = Buffer.from(base64Tar, "base64");
-  await extractArchive(tarBuffer, targetDir);
-}
 
 /** Restore the optional Claude marketplace catalog. */
 export async function applyClaudeMarketplace(content: string): Promise<void> {
@@ -530,8 +419,9 @@ export async function applyClaudeVault(
   vaultDir: string,
   key: string,
   dryRun: boolean,
-  options: ClaudeSyncOptions = {},
+  config: AgentSyncConfig,
 ): Promise<void> {
+  const syncMarketplace = config.claudePlugins?.syncMarketplace ?? false;
   const claudeDir = join(vaultDir, "claude");
   const files = await readAgeFiles(claudeDir);
 
@@ -562,7 +452,7 @@ export async function applyClaudeVault(
       // A vault that contains a marketplace.json.age is silently ignored on
       // any machine where syncMarketplace is not set, so opting out is the
       // safe default and never blocks pulls.
-      if (options.syncMarketplace !== true) {
+      if (!syncMarketplace) {
         continue;
       }
       if (dryRun) {
@@ -648,126 +538,4 @@ export async function applyClaudeVault(
   //   claude/plugins/<plugin>/mcp.json.age
   //   claude/plugins/<plugin>/skills/<skill>.tar.age
   await applyClaudePluginsDir(join(claudeDir, "plugins"), key, dryRun);
-}
-
-/**
- * Walk the `claude/plugins/` vault sub-tree and route each artifact to the
- * matching apply helper. Plugin names from the vault are validated via
- * {@link validatePluginName} before any filesystem write so a crafted vault
- * cannot escape `~/.claude/plugins/`.
- */
-async function applyClaudePluginsDir(
-  pluginsVaultDir: string,
-  key: string,
-  dryRun: boolean,
-): Promise<void> {
-  let pluginEntries: string[];
-  try {
-    pluginEntries = await readdir(pluginsVaultDir);
-  } catch {
-    return;
-  }
-
-  for (const pluginName of pluginEntries) {
-    // Symmetric with the push-side walker (claude-plugins.ts): dot-prefixed
-    // entries are ignorable noise (.gitkeep, .DS_Store), not adversarial. Skip
-    // silently before validation so the warning channel stays reserved for
-    // genuinely hostile names like `..`.
-    if (pluginName.startsWith(".")) continue;
-    try {
-      validatePluginName(pluginName);
-    } catch (err) {
-      if (err instanceof InvalidPluginNameError) {
-        log.warn(`[claude] Skipping vault plugin with invalid name '${pluginName}': ${err.reason}`);
-        continue;
-      }
-      throw err;
-    }
-
-    const pluginVaultRoot = join(pluginsVaultDir, pluginName);
-
-    // Top-level plugin artifacts (manifest, mcp.json).
-    const topFiles = await readAgeFiles(pluginVaultRoot);
-    for (const { name, fullPath } of topFiles) {
-      const encrypted = await readFile(fullPath, "utf8");
-      const decrypted = await decryptString(encrypted, key);
-
-      if (name === "plugin.json.age") {
-        if (dryRun) {
-          log.info(`[dry-run] [claude] would apply plugin manifest: ${pluginName}`);
-        } else {
-          await applyClaudePluginManifest(pluginName, decrypted);
-        }
-      } else if (name === "mcp.json.age") {
-        if (dryRun) {
-          log.info(`[dry-run] [claude] would apply plugin .mcp.json: ${pluginName}`);
-        } else {
-          await applyClaudePluginMcp(pluginName, decrypted);
-        }
-      }
-    }
-
-    // commands/*.md.age
-    for (const { name, fullPath } of await readAgeFiles(join(pluginVaultRoot, "commands"))) {
-      if (!name.endsWith(".md.age")) continue;
-      const encrypted = await readFile(fullPath, "utf8");
-      const decrypted = await decryptString(encrypted, key);
-      const fileName = basename(name, ".age");
-      if (dryRun) {
-        log.info(`[dry-run] [claude] would write plugin command: ${pluginName}/${fileName}`);
-      } else {
-        await applyClaudePluginCommand(pluginName, fileName, decrypted);
-      }
-    }
-
-    // agents/*.md.age
-    for (const { name, fullPath } of await readAgeFiles(join(pluginVaultRoot, "agents"))) {
-      if (!name.endsWith(".md.age")) continue;
-      const encrypted = await readFile(fullPath, "utf8");
-      const decrypted = await decryptString(encrypted, key);
-      const fileName = basename(name, ".age");
-      if (dryRun) {
-        log.info(`[dry-run] [claude] would write plugin agent: ${pluginName}/${fileName}`);
-      } else {
-        await applyClaudePluginAgent(pluginName, fileName, decrypted);
-      }
-    }
-
-    // hooks/*.json.age
-    for (const { name, fullPath } of await readAgeFiles(join(pluginVaultRoot, "hooks"))) {
-      if (!name.endsWith(".json.age")) continue;
-      const encrypted = await readFile(fullPath, "utf8");
-      const decrypted = await decryptString(encrypted, key);
-      const fileName = basename(name, ".age");
-      if (dryRun) {
-        log.info(`[dry-run] [claude] would write plugin hook: ${pluginName}/${fileName}`);
-      } else {
-        await applyClaudePluginHook(pluginName, fileName, decrypted);
-      }
-    }
-
-    // skills/<skill>.tar.age
-    for (const { name, fullPath } of await readAgeFiles(join(pluginVaultRoot, "skills"))) {
-      if (!name.endsWith(".tar.age")) continue;
-      const skillName = basename(name, ".tar.age");
-      try {
-        validateSkillName(skillName);
-      } catch (err) {
-        if (err instanceof InvalidSkillNameError) {
-          log.warn(
-            `[claude] Skipping plugin '${pluginName}' skill with invalid name '${name}': ${err.reason}`,
-          );
-          continue;
-        }
-        throw err;
-      }
-      const encrypted = await readFile(fullPath, "utf8");
-      const decrypted = await decryptString(encrypted, key);
-      if (dryRun) {
-        log.info(`[dry-run] [claude] would extract plugin skill: ${pluginName}/${skillName}`);
-        continue;
-      }
-      await applyClaudePluginSkill(pluginName, skillName, decrypted);
-    }
-  }
 }
