+++ src/core/__tests__/sanitizer.test.ts
@@ -5,10 +5,7 @@ import {
   NEVER_SYNC_PATTERNS,
   redactionEnvNameForPath,
   redactSecretLiterals,
-  sanitizeClaudeHooks,
-  sanitizeClaudeMcp,
-  sanitizeClaudePluginManifest,
-  sanitizeClaudePluginMcp,
+  sanitizeAndNormalizeJson,
   scanForSecrets,
   shouldNeverSync,
 } from "../sanitizer";
@@ -21,25 +18,6 @@ describe("sanitizer", () => {
     const value = result.value as { token: string };
     expect(value.token.startsWith("$AGENTSYNC_REDACTED")).toBeTrue();
     expect(result.warnings.length).toBe(1);
-  });
-
-  test("extracts hooks only from claude settings", () => {
-    const output = sanitizeClaudeHooks(
-      JSON.stringify({
-        hooks: { PreToolUse: [] },
-        other: { should: "be-dropped" },
-      }),
-    );
-    const parsed = JSON.parse(output.value) as Record<string, unknown>;
-    expect(Object.keys(parsed)).toEqual(["hooks"]);
-  });
-
-  test("extracts mcpServers only from .claude.json", () => {
-    const output = sanitizeClaudeMcp(
-      JSON.stringify({ mcpServers: { test: { command: "npx" } }, other: 123 }),
-    );
-    const parsed = JSON.parse(output.value) as Record<string, unknown>;
-    expect(Object.keys(parsed)).toEqual(["mcpServers"]);
   });
 
   test("never sync excludes sensitive files", () => {
@@ -151,53 +129,6 @@ describe("sanitizer", () => {
     expect(redactSecretLiterals(true).warnings).toHaveLength(0);
   });
 
-  // Claude plugin sanitizers (issue #31)
-
-  test("sanitizeClaudePluginManifest preserves manifest metadata", () => {
-    const manifest = JSON.stringify({
-      name: "acme-toolkit",
-      version: "1.2.3",
-      description: "Acme tools",
-      author: "Acme",
-      commands: [{ name: "review" }],
-    });
-    const out = sanitizeClaudePluginManifest(manifest);
-    const parsed = JSON.parse(out.value) as Record<string, unknown>;
-    expect(parsed.name).toBe("acme-toolkit");
-    expect(parsed.version).toBe("1.2.3");
-    expect(parsed.description).toBe("Acme tools");
-    expect((parsed.commands as { name: string }[])[0]?.name).toBe("review");
-  });
-
-  test("sanitizeClaudePluginManifest redacts literal secrets", () => {
-    const manifest = JSON.stringify({
-      name: "acme",
-      env: { token: `sk-${"x".repeat(30)}` },
-    });
-    const out = sanitizeClaudePluginManifest(manifest);
-    const parsed = JSON.parse(out.value) as { env: { token: string } };
-    expect(parsed.env.token.startsWith("$AGENTSYNC_REDACTED")).toBeTrue();
-    expect(out.warnings.length).toBe(1);
-  });
-
-  test("sanitizeClaudePluginMcp preserves the bare server descriptor", () => {
-    const mcp = JSON.stringify({
-      mcpServers: { acme: { command: "node", args: ["server.js"] } },
-    });
-    const out = sanitizeClaudePluginMcp(mcp);
-    const parsed = JSON.parse(out.value) as Record<string, unknown>;
-    expect(parsed.mcpServers).toBeDefined();
-  });
-
-  test("sanitizeClaudePluginMcp redacts literal secrets inside server config", () => {
-    const mcp = JSON.stringify({
-      mcpServers: { acme: { env: { API_KEY: `sk-${"y".repeat(30)}` } } },
-    });
-    const out = sanitizeClaudePluginMcp(mcp);
-    expect(out.warnings.length).toBe(1);
-    expect(out.value).toContain("$AGENTSYNC_REDACTED");
-  });
-
   test("EMBEDDED_SECRET_PATTERNS is non-empty so scanForSecrets has rules to run", () => {
     expect(EMBEDDED_SECRET_PATTERNS.length).toBeGreaterThan(0);
     for (const entry of EMBEDDED_SECRET_PATTERNS) {
@@ -269,63 +200,25 @@ describe("sanitizer", () => {
     expect(scanForSecrets(prose, "/tmp/notes.md")).toEqual([]);
   });
 
-  // HOME path portability wiring. Each sanitize function must rewrite a literal
-  // home-prefix to the ${AGENTSYNC_HOME} placeholder before serializing so the
-  // vault round-trips across machines with different home directories.
-
-  test("sanitizeClaudeHooks rewrites home-prefixed paths to placeholder", () => {
-    const home = "/home/alpha";
-    const raw = JSON.stringify({
-      hooks: { PreToolUse: [{ command: `${home}/.claude/runner` }] },
-    });
-    const out = JSON.parse(sanitizeClaudeHooks(raw, home).value) as {
-      hooks: { PreToolUse: { command: string }[] };
-    };
-    expect(out.hooks.PreToolUse[0]?.command).toBe(`${AGENTSYNC_HOME_PLACEHOLDER}/.claude/runner`);
-  });
-
-  test("sanitizeClaudeMcp rewrites home-prefixed paths to placeholder", () => {
+  test("sanitizeAndNormalizeJson rewrites home paths and redacts literal secrets", () => {
     const home = "/Users/alpha";
     const raw = JSON.stringify({
-      mcpServers: { fs: { command: "node", cwd: `${home}/proj` } },
+      cwd: `${home}/proj`,
+      env: { TOKEN: "sk-AAAAAAAAAAAAAAAAAAAAAA" },
     });
-    const out = JSON.parse(sanitizeClaudeMcp(raw, home).value) as {
-      mcpServers: { fs: { cwd: string } };
-    };
-    expect(out.mcpServers.fs.cwd).toBe(`${AGENTSYNC_HOME_PLACEHOLDER}/proj`);
+    const out = sanitizeAndNormalizeJson(raw, "test", home);
+    const parsed = JSON.parse(out.value) as { cwd: string; env: { TOKEN: string } };
+    expect(parsed.cwd).toBe(`${AGENTSYNC_HOME_PLACEHOLDER}/proj`);
+    expect(parsed.env.TOKEN).not.toBe("sk-AAAAAAAAAAAAAAAAAAAAAA");
+    expect(out.warnings.length).toBeGreaterThan(0);
   });
 
-  test("sanitizeClaudePluginManifest rewrites home-prefixed paths to placeholder", () => {
-    const home = "/Users/alpha";
-    const raw = JSON.stringify({
-      name: "lefthook",
-      installPath: `${home}/.claude/plugins/lefthook`,
-    });
-    const out = JSON.parse(sanitizeClaudePluginManifest(raw, home).value) as {
-      installPath: string;
-      name: string;
-    };
-    expect(out.installPath).toBe(`${AGENTSYNC_HOME_PLACEHOLDER}/.claude/plugins/lefthook`);
-    expect(out.name).toBe("lefthook");
+  test("sanitizeAndNormalizeJson throws on invalid JSON without silent fallback", () => {
+    expect(() => sanitizeAndNormalizeJson("not json", "test")).toThrow();
   });
 
-  test("sanitizeClaudePluginMcp rewrites home-prefixed paths to placeholder", () => {
-    const home = "/Users/alpha";
-    const raw = JSON.stringify({
-      command: "node",
-      args: [`${home}/.claude/plugins/x/srv.js`, "--root=/etc/hosts"],
-    });
-    const out = JSON.parse(sanitizeClaudePluginMcp(raw, home).value) as { args: string[] };
-    expect(out.args[0]).toBe(`${AGENTSYNC_HOME_PLACEHOLDER}/.claude/plugins/x/srv.js`);
-    expect(out.args[1]).toBe("--root=/etc/hosts");
-  });
-
-  test("sanitize functions pass through unchanged when home is empty", () => {
-    const home = "/Users/alpha";
-    const raw = JSON.stringify({ hooks: { PreToolUse: [{ command: `${home}/runner` }] } });
-    const out = JSON.parse(sanitizeClaudeHooks(raw, "").value) as {
-      hooks: { PreToolUse: { command: string }[] };
-    };
-    expect(out.hooks.PreToolUse[0]?.command).toBe(`${home}/runner`);
+  test("sanitizeAndNormalizeJson appends trailing newline for diff stability", () => {
+    const out = sanitizeAndNormalizeJson(JSON.stringify({ a: 1 }), "test", "/home/x");
+    expect(out.value.endsWith("\n")).toBe(true);
   });
 });
