+++ src/migrate/translators/frontmatter.ts
@@ -1,18 +1,21 @@
 /**
- * src/migrate/translators/_frontmatter.ts
+ * Internal to `src/migrate/translators`.
  *
  * Minimal YAML frontmatter parser/serializer shared by skills, rules,
  * and commands-to-skills translators. Handles the small subset of YAML
  * that AI agents use in their config frontmatter:
  *   - Top-level scalar key/value pairs
  *   - Quoted and unquoted string values
- *   - Boolean values (true/false)
- *   - Comma-separated lists (kept as raw string)
+ *   - Boolean values true/false
+ *   - Comma-separated lists, kept as raw string
  *
- * Not a full YAML parser by design — pulling in `js-yaml` for what the
- * spec actually uses would be overkill and adds a dependency. If the
+ * Not a full YAML parser by design — pulling in `js-yaml` for what these
+ * configs actually use would be overkill and adds a dependency. If the
  * frontmatter has nested structures, we fall back to preserving the
  * raw block as-is.
+ *
+ * Used by: commands.ts, rules.ts, skills.ts.
+ * Not used by: global-rules.ts, mcp.ts.
  */
 
 export interface Frontmatter {
